backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
azure-core==1.34.0
azure-storage-blob==12.25.1
azure-storage-file-datalake==12.20.0
cachetools==6.0.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.3
cssselect==1.3.0
cssutils==2.11.1
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
environs==14.2.0
fastapi==0.115.12
fastapi_cors==0.0.6
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
isodate==0.7.2
jiter==0.10.0
lxml==5.4.0
marshmallow==4.0.0
more-itertools==10.7.0
openai==1.84.0
premailer==3.10.0
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pymongo==4.13.0
python-dotenv==1.1.0
python-multipart==0.0.20
requests==2.32.3
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
uvicorn==0.34.3
yagmail==0.15.293
